added new test format

+++ ProgramTwo/src/main.cpp
@@ -98,8 +98,8 @@ void printMST(Graph* graph, std::vector<Edge*> edges) {
  * @param edges the edges of the MST
  * @return long the total weight of the MST
  */
-long totalWeight(vector<Edge*> edges) {
-    long sum = 0;
+long long totalWeight(vector<Edge*> edges) {
+    long long sum = 0;
     for (auto e : edges) {
         sum += e->weight;
     }
